+++ src/vba_edit/excel_vba.py
@@ -74,9 +74,7 @@ IMPORTANT:
 
     # Add --version argument to the main parser
     parser.add_argument(
-        "--version",
-        action="version",
-        version=f"{package_name_formatted} v{package_version} ({entry_point_name})"
+        "--version", action="version", version=f"{package_name_formatted} v{package_version} ({entry_point_name})"
     )
 
     subparsers = parser.add_subparsers(dest="command", required=True)
@@ -104,7 +102,10 @@ IMPORTANT:
                 "help": "Use wrapper for xlwings vba edit|import|export commands",
             },
         ),
-        "version": (["--version"], {"action": "version", "version": f"{package_name_formatted} v{package_version} ({entry_point_name})"}),
+        "version": (
+            ["--version"],
+            {"action": "version", "version": f"{package_name_formatted} v{package_version} ({entry_point_name})"},
+        ),
     }
 
     # Edit command