+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for macro-enabled MS PowerPoint documents
 - Option to show program's version number and exit added to all cli interfaces (`--version`)
 <!-- ### Changed -->
 <!-- -  -->